fix: cast size_t widths to int for variadic printer to avoid Windows x64 crashes

+++ src/console_reporter.cc
@@ -137,7 +137,7 @@ void ConsoleReporter::PrintRunData(const Run& result) {
                            : IgnoreColorPrint;
   auto name_color =
       (result.report_big_o || result.report_rms) ? COLOR_BLUE : COLOR_GREEN;
-  printer(Out, name_color, "%-*s ", name_field_width_,
+  printer(Out, name_color, "%-*s ", static_cast<int>(name_field_width_),
           result.benchmark_name().c_str());
 
   if (internal::SkippedWithError == result.skipped) {
@@ -196,8 +196,8 @@ void ConsoleReporter::PrintRunData(const Run& result) {
       }
     }
     if ((output_options_ & OO_Tabular) != 0) {
-      printer(Out, COLOR_DEFAULT, " %*s%s", cNameLen - strlen(unit), s.c_str(),
-              unit);
+      printer(Out, COLOR_DEFAULT, " %*s%s",
+              static_cast<int>(cNameLen - strlen(unit)), s.c_str(), unit);
     } else {
       printer(Out, COLOR_DEFAULT, " %s=%s%s", c.first.c_str(), s.c_str(), unit);
     }
